Add shorthand createThemeHook function
Purpose is to avoid having to pass in styles explicitly as an argument
